+++ client/src/pages/EditUser.jsx
@@ -1,128 +1,49 @@
-import React, { useEffect, useState } from 'react';
-// import { CgOptions } from 'react-icons/cg';
+import React, {useEffect, useState} from 'react';
+import s from './styles/EditUser.module.css';
+import {useGlobalStorage} from '../hooks/useGlobalStorage';
 import { useDispatch, useSelector } from "react-redux";
-import { getByUserId, filterProfessions } from '../redux/actions'
+import { getByUserId, filterProfessions } from '../redux/actions';
 import axios from 'axios';
-import s from './styles/EditPodeddional.module.css'
-import {useGlobalStorage} from '../hooks/useGlobalStorage'
-import { CgOptions } from 'react-icons/cg';
 
+export default function EditCliente() {
 
-export default function EditUser() {
-    
     const [globalUser, setGlobalUser] = useGlobalStorage("globalUser", "");
-    const[errors, setErrors] = useState({});
+    const login = !localStorage.getItem ? null: JSON.parse(localStorage.getItem("user"))
     const dispatch = useDispatch();
+    const[errors, setErrors] = useState({});
+    const [buttonSubmit, setbuttonSubmit] = useState(false)
     const oficio = useSelector((state) => state.professionsName)
-    // console.log(oficio)
-    const user = useSelector((state) => state.globalUserGlobalStorage);
-    const state = useSelector((state) => state);
-    // const login = !localStorage.getItem ? null: JSON.parse(localStorage.getItem("user"))
-    
-    
-    console.log('user linea20', globalUser)
-    
-    
-    useEffect(() => {
-        dispatch(getByUserId(user.id))
-        dispatch(filterProfessions (oficio))
-        
-    },[])
-
-    // console.log('prfesional: ', user[0]?.Professional?.Professions[0]?.name)
     const [profession, setProfession] = useState([])
-    console.log(profession)
+
     const [details, setDetails] = useState({
-        firstName:user[0]?.first_name,
-        lastName: user[0]?.last_name,
-        email: user[0]?.email,
-        dni:user[0]?.dni,
-        password:user[0]?.password,
-        repeatPassword:user[0]?.password,
-        // professional:user[0]?.professional,
-        // cliente:'',
+        firstName: globalUser.first_name,
+        lastName: globalUser.last_name,
+        email: globalUser.email,
+        professional: globalUser.professional,
         professionalCase:false,
-        city:user[0]?.city,
         profession:profession.toString(),
-        phone:'',
-        photo:'', 
     })
-    console.log(details)
+
+    console.log('profecional: ', globalUser.professional)
+    useEffect(() => {
+        dispatch(getByUserId(login?.id))  
+        dispatch(filterProfessions (oficio))      
+    },[])
+    
+    
     function validate(valores){
         let errores = {};
         
         //validacion nombre
-        // console.log('error: ',errores)
+        console.log('error: ',valores)
         if(!valores.firstName) {
             errores.firstName = 'Por favor ingresa tu nombre'
         }else if(!/^[a-zA-ZÀ-ÿ\s]{1,40}$/.test(valores.firstName)){
             errores.firstName= 'El nombre solo puede contener letras y espacios'
         }
-    
-        //validar apellido 
-        if(!valores.lastName) {
-            errores.lastName = 'Por favor ingresa tu apellido'
-        }else if(!/^[a-zA-ZÀ-ÿ\s]{1,40}$/.test(valores.lastName)){
-            errores.lastName= 'El apellido solo puede contener letras y espacios'
-        }
-    
-        //validacion correo
-        if(!valores.email) {
-            errores.email = 'Por favor ingresa tu correo electronico'
-        }else if(! /^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$/.test(valores.email)){
-            errores.email= 'El correo solo puede contener letras,numeros, puntos, guiones y guion bajo'
-        }
-    
-        //validacion password
-        if(!valores.password) {
-            errores.password = 'Por favor ingresa un password'
-        }else if(! /^(?=.[A-Za-z])(?=.\d)[A-Za-z\d]{8,}$/.test(valores.password)){
-            errores.password= 'El password debe tener mínimo ocho caracteres, al menos una letra y un número'
-        }
-    
-        //validacion Repeat-password
-        if(!(valores.password === valores.repeatPassword)) {
-            errores.repeatPassword = 'Por favor la password tiene que ser igual'
-        }
-
-        //validacion télefono /^(?:(?:00)?549?)?0?(?:11|[2368]\d)(?:(?=\d{0,2}15)\d{2})??\d{8}$/
-        if (!valores.phone) {
-            errores.phone = 'Por favor ingresa un télefono'
-        }else if(!/^(?:(?:00)?549?)?0?(?:11|[2368]\d)(?:(?=\d{0,2}15)\d{2})??\d{8}$/.test(valores.phone)){
-            errors.phone= 'Es obligatorio: el código de área (11, 2xx, 2xxx, 3xx, 3xxx, 6xx y 8xx), (no toma como válido un número local sin código de área como 4444-0000)'
-        }
 
         return errores;
     }
-
-    function handleSelect(e){
-        console.log('profecion:', profession)
-        setDetails({
-            ...details,
-            profession: details.profession + e.target.id + ","
-        })
-    }
-    function onClose(e){
-        let index = profession.indexOf(e.target.value)
-        setProfession([...profession.slice(0, index).concat(...profession.slice(index+1, profession.length))])
-        setDetails({
-            ...details,
-            profession: details.profession + e.target.id + ","
-        })
-    }
-
-    function onClick(e){
-        if(profession.indexOf(details.profession) === -1 && details.profession !== ''){
-        setProfession([...profession, details.profession])
-        // setDetails({...details, profession : 'Selecciona a um oficio'})
-        }
-        // setDetails({...details, profession : 'Selecciona a um oficio'})
-
-    }
-
-    function changeCountry(event){
-        setDetails({...details, profession: event.target.value})
-    }   
 
     const handleSubmit = async (e) =>{
         e.preventDefault();
@@ -130,79 +51,49 @@ export default function EditUser() {
         const err = validate(details)
 
         setErrors(err)
-        
-        // const id = user[0].id
 
-    try{
-        let prof= profession.toString()
-        console.log(prof)
-        let newData= {
-            firstName:details.firstName,
-            lastName: details.lastName,
-            email: details.email,
-            dni:details.dni,
-            password:details.password,
-            // professional:user[0]?.professional,
-            // cliente:'',
-            city:details.city,
-            profession:prof,
-            phone:'',
-            photo:'', 
+        try{
+            let prof= profession.toString()
+            let newData= {
+                firstName:details.firstName,
+                lastName: details.lastName,
+                email: details.email,
+                dni:details.dni,
+                password:details.password,
+                professional: details.professional,
+                profession:prof,
+
+            }
+            await axios.put(`http://localhost:3001/user/updateUser/101`, newData)
+            setGlobalUser({
+                ...globalUser,
+                first_name:details.firstName,
+                last_name: details.lastName,
+                email: details.email,
+                dni:details.dni,
+                password:details.password,
+                rofessional: details.professional,
+                
+            })
+        }catch(e){
+            console.log(e)
         }
-        await axios.put(`http://localhost:3001/user/updateUser/101`, newData)
-        setGlobalUser({
-            ...globalUser,
-            first_name:details.firstName,
-            last_name: details.lastName,
-            email: details.email,
-            dni:details.dni,
-            password:details.password,
-            // professional:user[0]?.professional,
-            city:details.city, 
-        })
-        console.log('combio')
-    }catch(e){
-        console.log(e)
     }
-    
-    
-    // console.log('id: ',id)
-    }
-    
-    useEffect(() => {
-        setDetails({
-            ...details,
-            firstName: state.user[0]?.first_name,
-            lastName: user[0]?.last_name,
-            email: user[0]?.email,
-            dni:user[0]?.dni,
-            password:user[0]?.password,
-            repeatPassword:user[0]?.password,
-            professional:user[0]?.professional,
-            cliente:'',
-            city:user[0]?.city,
-            profession:  user[0]?.Professional?.Professions[0]?.name,
-            phone:'',
-            photo:'', 
-
-        })
-    
-    }, [state])
-
-    useEffect(() => {
-        setErrors(validate(details))    
-    }, [details])
 
     function handleChange(e){
         setDetails({
             ...details,
             [e.target.name] : e.target.value
+            
         })
-        // console.log('detalle: ',details)
+        setErrors(
+            validate(errors)
+        )
+        
     }
+
     function handleCheck(e){
-        console.log("checked204",e.target.checked)
-        console.log("checked-value",e.target.value)
+        console.log(details.professionalCase)
         if (e.target.id === 'checkboxClient') {
             setDetails({
                 ...details,
@@ -218,6 +109,24 @@ export default function EditUser() {
             })
         }
     } 
+
+    function onClick(e){
+        if(profession.indexOf(details.profession) === -1 && details.profession !== ''){
+        setProfession([...profession, details.profession])
+        }
+    }
+    function changeCountry(event){
+        setDetails({...details, profession: event.target.value})
+    }  
+    function onClose(e){
+        let index = profession.indexOf(e.target.value)
+        setProfession([...profession.slice(0, index).concat(...profession.slice(index+1, profession.length))])
+        setDetails({
+            ...details,
+            profession: details.profession + e.target.id + ","
+        })
+    }
+    
     return (
         <div className={s.container}>
             <div className={ s.container_img}>
@@ -229,7 +138,6 @@ export default function EditUser() {
                 <div className={s.container_edilt_titulo}>
                     <h2>Edita tu perfil</h2>
                 </div>
-
             <form className={s.container_edilt_form} onSubmit={(e) => handleSubmit(e)}>
                 <div className={s.container_edilt_form_input}>
                     <label >Nombre: </label>
@@ -248,24 +156,24 @@ export default function EditUser() {
                     <input 
                         type="text"  
                         name="lastName"
-                        value={details.lastName}
+                        value={globalUser.last_name}
                         onChange={(e) => handleChange(e)}
                     />
-                    {!errors.name && (
+                    {/* {!errors.name && (
                         <p >{errors.lastName}</p>
-                    )}
+                    )} */}
                 </div>
                 <div className={s.container_edilt_form_input}>
                     <label >Correo: </label>
                     <input 
                         type='email' 
                         name='email'
-                        value={details.email}
+                        value={globalUser.email}
                         onChange={(e) => handleChange(e)}
                     />
-                    {errors.email && (
+                    {/* {errors.email && (
                         <p>{errors.email}</p>
-                    )}
+                    )} */}
                 </div>
                 <div className={s.container_edilt_form_input}>
                     <label>DNI:</label>
@@ -273,7 +181,6 @@ export default function EditUser() {
                         type='text'
                         name='dni'
                         readonly
-                        value={details.dni}
                         />
                 </div>
                 <div className={s.container_edilt_form_input}>
@@ -281,39 +188,36 @@ export default function EditUser() {
                     <input 
                         type='password'
                         name='password'
-                        value={details.password}
-                        onChange={(e) => handleChange(e)}
+                        // onChange={(e) => handleChange(e)}
                     />
-                    {errors.password && (
+                    {/* {errors.password && (
                         <p>{errors.password}</p>
-                    )}
+                    )} */}
                 </div>
                 <div className={s.container_edilt_form_input}>
                     <label>Repeat Password:</label>
                     <input 
                         type='password'
                         name='repeatPassword'
-                        value={details.repeatPassword}
-                        onChange={(e) => handleChange(e)}
+                        // onChange={(e) => handleChange(e)}
                     />
-                    {errors.repeatPassword && (
+                    {/* {errors.repeatPassword && (
                         <p>{errors.repeatPassword}</p>
-                    )}
+                    )} */}
                 </div>
                 <div className={s.container_edilt_form_input}>
                     <label >Teléfono: </label>
                     <input 
                         type="text"  
                         name="phone"
-                        value={details.phone}
-                        onChange={(e) => handleChange(e)}
-                        />
-                    {errors.photo && (
+                        // onChange={(e) => handleChange(e)}
+                    />
+                    {/* {errors.photo && (
                         <p>{errors.photo}</p>
-                        )}
+                    )} */}
                 </div>
 
-                <div >
+                <div className={s.container_edilt_form_input_img} >
                     <label for="imageFile">Selecciona alguna imágen (png):</label><br/>
                     <div className={s.div_file}>
                         <p className={s.text}>Elegir archivo</p>
@@ -321,18 +225,16 @@ export default function EditUser() {
                             type="file"  
                             accept=".png" 
                             multiple
-                            value={details.photo}
-                            onChange={(e) => handleChange(e)}
+                            // onChange={(e) => handleChange(e)}
                         /> 
                     </div>
                 </div>
-                    
+
                 {
                     globalUser.professional ?
                     
-////----------------------- edit profesional--------------------                    
-                
-                    <div>
+////////----------------------- edit Profesional -------------------- 
+
                     <div className={s.subDiv}>
                         <label for="countries">Seleccona tu Oficio:</label>
                         <div>
@@ -357,90 +259,6 @@ export default function EditUser() {
                         <div className={s.countriesDiv}>
 
                             {
-                                profession.map(e => {
-                                    return(
-                                        <input type="button" value={e} className={s.countryBtn}  onClick={onClose} />
-                                        )
-                                        
-                                    })}
-                        </div>
-                    </div>
-                    </div>
-
-                    :
-////----------------------------- edit cliente ------------------------------
-                    <div className={s.container_edilt_form_pofcional}>
-                        <label>Dar de alta como profesional:<input
-                                type='checkbox'
-                                id="checkbox"
-                                value='professional'
-                                checked={details.professionalCase}
-                                onChange={(e) => handleCheck(e)}
-                            />
-                        </label>
-                        <label>Seguir registrado como cliente:<input
-                            type='checkbox'
-                            id="checkboxClient"
-                            checked=""
-                            value='cliente'
-                            onChange={(e) => handleCheck(e)}
-                            />
-                        </label>
-                    <>
-                    <div className={s.container_registro_form_check}>
-                    </div>
-                    <div className={s.container_registro_form_oficio}>
-                        {
-                            details.professionalCase ? (
-                                <>
-                                    <label>Seleccona tu Oficio:</label>
-                                    <div className='dropdown'>
-                                        <button className="btn btn-secondary dropdown-toggle" id="dropdownMenuButton1" data-bs-toggle="dropdown" type="button" aria-expanded="false" ><CgOptions />Elegir oficio</button>
-                                        <ul className="dropdown-menu" aria-labelledby="dropdownMenuButton1" >
-                                            {
-                                                oficio.map((el, index) => {
-                                                    return (
-                                                        <li key={"li" + index}><span className="dropdown-item" id={el} onClick={handleSelect}>{el}</span></li>
-                                                    )
-                                                })
-                                            }
-                                        </ul>
-                                    </div>
-                                </>
-                            ) : (<></>)
-                        }
-
-                    </div>
-
-
-                    </>
-                    
-                    
-
-                    {/* <div className={s.subDiv}>
-                        <label for="countries">Seleccona tu Oficio:</label>
-                        <div>
-                            <select 
-                            className={s.inputClass3}
-                            onChange={changeCountry}
-                            value={details.profession}
-                            >
-                            <option >Elige tu profesión</option>
-                            {oficio.map(e => {
-                                return (<option >{e}</option>)
-                            })
-                            }
-                            </select>
-                            <input 
-                            onClick={onClick} 
-                            className={s.btnAdd} 
-                            type="button" 
-                            value="+" 
-                            />
-                        </div>
-                        <div className={s.countriesDiv}>
-
-                            {
                             profession.map(e => {
                                 return(
                                 <input type="button" value={e} className={s.countryBtn}  onClick={onClose} />
@@ -448,20 +266,73 @@ export default function EditUser() {
                                 
                                 })}
                         </div>
-                    </div> */}
+                        <div>
+                            <label>Dar de baja como profesional:<input
+                            type='checkbox'
+                            id="checkboxClient"
+                            checked={!details.professionalCase}
+                            value='cliente'
+                            onChange={(e) => handleCheck(e)}
+                        /></label>    
+                        </div>
+                    </div>
+
+                    :
+
+//////--------------------- edit de cliente -------------------------------
+
+                    <div className={s.container_edilt_form_pofcional}>
+                        <label>Dar de alta como profecional:<input
+                                type='checkbox'
+                                value={globalUser.professional}
+                                checked={details.professionalCase}
+                                onChange={(e) => handleCheck(e)}
+                        /></label>
+                        <div className={s.subDiv}>
+                            <label for="countries">Seleccona tu Oficio:</label>
+                            <div>
+                                <select 
+                                className={s.inputClass3}
+                                onChange={changeCountry}
+                                value={details.profession}
+                                >
+                                <option >Elige tu profesión</option>
+                                {oficio.map(e => {
+                                    return (<option >{e}</option>)
+                                })
+                                }
+                                </select>
+                                <input 
+                                onClick={onClick} 
+                                className={s.btnAdd} 
+                                type="button" 
+                                value="+" 
+                                />
+                            </div>
+                            <div className={s.countriesDiv}>
+
+                                {
+                                profession.map(e => {
+                                    return(
+                                    <input type="button" value={e} className={s.countryBtn}  onClick={onClose} />
+                                    )
+                                    
+                                    })}
+                            </div>
+                        </div>
+                        
+                    
                 </div>
+
                 }
 
-                    <div >
-                        <p>Recuerda que los campos que no edites no se van a cambiar, lo unico que no podrás editar es el DNI.</p>  
-                    </div>
+                <div >
+                    <p>Reguerdas que los campos que no edites nada no se  van a cambian, lo unico que no se 
+                    va a poder editar es el DNI.</p>  
+                </div>
                 
-                <button type='submit' className={"btn btn-success " + s.buttonSubmit} >Cambiar Perfil</button>
+                <button type='submit'id='buttonSubmit' className={"btn btn-success " + s.buttonSubmit} >Cambiar Perfil</button>
             </form>
-
-            
-
-
             </div>
         </div>
     )
