--- client/src/pages/styles/EditPodeddional.module.css
@@ -1,175 +1,0 @@
-.btn{
-    background-color:#233142;
-    color: white;
-    width: 285px;
-    height: 40px;
-    border-radius:5px;
-    cursor:pointer;
-    margin:10px;
-    cursor:url("https://img.icons8.com/color/50/000000/airplane-take-off--v2.png"), auto;
-}
-
-.btnAdd{
-    height:32px;
-    background-color:#233142;
-    color:white;
-    cursor:pointer;
-}
-.btnBlock{
-    background-color:grey;
-    color: white;
-    width: 285px;
-    height: 40px;
-    border-radius:5px;
-    cursor: not-allowed;
-    margin:10px;
-
-}
-.countriesDiv{
-    margin-top:10px;
-    display:flex;
-    background-color: white;
-    font-size:10px;
-    width:400px;
-    height:200px;
-    flex-direction: row;
-    flex-wrap: wrap;
-    align-content: flex-start;
-    border: #142d4c 2px solid;
-}
-.countryBtn{
-    width:25%;
-    height:20%;
-    white-space: normal;
-    font-size:11.5px;
-    border-radius:5px;
-    background-color:#2b4353;
-    color: white;
-    cursor:pointer;
-    font-weight: 700;
-}
-.countryBtn:hover{
-    background-color:#f95959;
-    color:black
-}
-.div{
-    display:flex;
-    height:93.5vh;
-    width:99.8%;
-    justify-content:center;
-}
-
-.form{
-    display:flex;
-    flex-direction: column;
-    align-items:center;
-}
-
-.inputClass{
-    background-color: white; 
-    color: black;
-    font-size: 13px;
-    border-radius:5px;
-    width: 270px;
-    height: 25px;
-}
-
-.inputClass2{
-    height:32px;
-    width:280px;
-    border-radius:5px;
-    border: black 2px solid;
-}
-
-.inputClass3{
-    height:32px;
-    width:256px;
-    border: black 1.6px solid;
-    border-radius:5px;
-}
-
-.subDiv{
-    display:flex;
-    flex-direction: column;
-    justify-content: space-between;
-    font-size:22px;
-    margin:10px;
-}
-
-.container{
-    display: flex;
-    height: 100vh; 
-}
-
-.container_edilt_titulo{
-    position: relative;
-    top: 20px;
-    width: 70vw;
-}
-
-.container_edilt_titulo h2{
-    text-align: center;
-}
-
-.container_img{
-    background-color: aqua;
-    display: grid;
-    align-content: space-around;
-    width: 30vw;
-}
-
-.container_img p{
-    color: #505050;
-    font-size: 25px;
-    font-weight: 100;
-    padding: 2em;
-}
-
-.container__edilt{
-    width: 70vw;
-}
-
-.container_edilt_form{
-    display: flex;
-    flex-wrap: wrap;
-    position: relative;
-    top: -10px;
-    padding: 50px;
-}
-
-.container_edilt_form_input{
-    width: 30%;
-    margin: 10px;
-}
-
-.buttonSubmit{
-    display: inline-block;
-    width: 400px;
-    max-height: 45px;
-    margin: auto;
-}
-
-.div_file{
-    position: relative;
-    padding: 8px;
-    margin: 15px;
-    width: 150px;
-    background-color: aqua;
-    border-radius: 5px;
-}
-.text{
-    text-align: center;
-    color: white;
-    margin: 0px;
-}
-
-.btn_enviar{
-    position: absolute;
-    top: 0px;
-    left: 0px;
-    right: 0px;
-    bottom: 0px;
-    width: 100%;
-    opacity: 0;
-
-}
